requirements.txt
praw>=7.7,<8
python-dotenv>=1.0,<2
